requirements.txt
hikari[speedups]==2.0.0.dev125
hikari-lightbulb==2.3.1
hikari-miru==3.0.2
python-dotenv==0.21.1